# src/api/v1/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nubabel Public API
  description: |
    Public API for Nubabel - Multi-Agent Orchestration Platform.

    ## Authentication
    All endpoints require Bearer token authentication using an API key.
    Include the header: `Authorization: Bearer <your_api_key>`

    ## Rate Limiting
    Rate limits depend on your API key tier:
    - **Free**: 60 requests/minute, 1,000 requests/day
    - **Pro**: 300 requests/minute, 10,000 requests/day
    - **Enterprise**: 1,000 requests/minute, 100,000 requests/day

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit-Minute`: Your per-minute limit
    - `X-RateLimit-Remaining-Minute`: Remaining requests this minute
    - `X-RateLimit-Limit-Day`: Your per-day limit
    - `X-RateLimit-Remaining-Day`: Remaining requests today

    ## Errors
    All errors follow a consistent format:
    ```json
    {
      "error": "error_code",
      "message": "Human-readable description"
    }
    ```
  version: 1.0.0
  contact:
    name: Nubabel Support
    email: support@nubabel.io
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: Production API
...
security:
  - BearerAuth: []
paths:
...
  /workflows:
    get:
      summary: List workflows
      tags: [Workflows]
      security:
        - BearerAuth: [workflows:read]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived, draft]
      responses:
        '200':
          description: List of workflows
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workflow'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      summary: Create workflow
      tags: [Workflows]
      security:
        - BearerAuth: [workflows:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflow'
      responses:
        '201':
          description: Workflow created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Workflow'
  /workflows/{workflowId}:
    get:
      summary: Get workflow details
      tags: [Workflows]
      security:
        - BearerAuth: [workflows:read]
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workflow details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/WorkflowDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update workflow
      tags: [Workflows]
      security:
        - BearerAuth: [workflows:write]
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkflow'
      responses:
        '200':
          description: Workflow updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Workflow'

    delete:
      summary: Delete workflow
      tags: [Workflows]
      security:
        - BearerAuth: [workflows:write]
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Workflow deleted
  /workflows/{workflowId}/execute:
    post:
      summary: Execute workflow
      tags: [Workflows]
      security:
        - BearerAuth: [workflows:execute]
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                input:
                  type: object
                  description: Input variables for the workflow
                sessionId:
                  type: string
      responses:
        '202':
          description: Workflow execution started
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      executionId:
                        type: string
                      workflowId:
                        type: string
                      status:
                        type: string
                      startedAt:
                        type: string
                        format: date-time
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from the Developer Portal
  schemas:
...
    Workflow:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, archived, draft]
        nodeCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkflowDetail:
      allOf:
        - $ref: '#/components/schemas/Workflow'
        - type: object
          properties:
            nodes:
              type: array
              items:
                type: object
            edges:
              type: array
              items:
                type: object
            triggerType:
              type: string
            lastExecutedAt:
              type: string
              format: date-time
    CreateWorkflow:
      type: object
      required:
        - name
        - nodes
      properties:
        name:
          type: string
        description:
          type: string
        nodes:
          type: array
          items:
            type: object
        edges:
          type: array
          items:
            type: object
        triggerType:
          type: string
          enum: [manual, webhook, schedule]
    UpdateWorkflow:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, archived, draft]
        nodes:
          type: array
          items:
            type: object
        edges:
          type: array
          items:
            type: object
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: not_found
              message:
                type: string
...